+++ packages/core/react/use-chat.ts
@@ -16,6 +16,7 @@ import type {
   ReactResponseRow,
   experimental_StreamingReactResponse,
 } from '../streams/streaming-react-response';
+
 export type { CreateMessage, Message, UseChatOptions };
 
 export type UseChatHelpers = {
@@ -73,6 +74,9 @@ export type UseChatHelpers = {
   data?: JSONValue[];
 };
 
+/**
+@deprecated Use AI SDK RSC instead: https://sdk.vercel.ai/docs/ai-sdk-rsc
+ */
 type StreamingReactResponseAction = (payload: {
   messages: Message[];
   data?: Record<string, string>;
@@ -101,20 +105,27 @@ const getStreamedResponse = async (
   const constructedMessagesPayload = sendExtraMessageFields
     ? chatRequest.messages
     : chatRequest.messages.map(
-        ({ role, content, name, function_call, tool_calls, tool_call_id }) => ({
+        ({
           role,
           content,
+          name,
+          toolInvocations,
+          function_call,
+          tool_calls,
           tool_call_id,
+        }) => ({
+          role,
+          content,
           ...(name !== undefined && { name }),
-          ...(function_call !== undefined && {
-            function_call: function_call,
-          }),
-          ...(tool_calls !== undefined && {
-            tool_calls: tool_calls,
-          }),
+          ...(toolInvocations !== undefined && { toolInvocations }),
+          // outdated function/tool call handling (TODO deprecate):
+          tool_call_id,
+          ...(function_call !== undefined && { function_call }),
+          ...(tool_calls !== undefined && { tool_calls }),
         }),
       );
 
+  // TODO deprecated, remove in next major release
   if (typeof api !== 'string') {
     // In this case, we are handling a Server Action. No complex mode handling needed.
 
@@ -208,6 +219,7 @@ export function useChat({
   sendExtraMessageFields,
   experimental_onFunctionCall,
   experimental_onToolCall,
+  experimental_maxAutomaticRoundtrips = 0,
   streamMode,
   onResponse,
   onFinish,
@@ -219,7 +231,28 @@ export function useChat({
 }: Omit<UseChatOptions, 'api'> & {
   api?: string | StreamingReactResponseAction;
   key?: string;
-} = {}): UseChatHelpers {
+  /**
+Maximal number of automatic roundtrips for tool calls.
+
+An automatic tool call roundtrip is a call to the server with the 
+tool call results when all tool calls in the last assistant 
+message have results.
+
+A maximum number is required to prevent infinite loops in the
+case of misconfigured tools.
+
+By default, it's set to 0, which will disable the feature.
+   */
+  experimental_maxAutomaticRoundtrips?: number;
+} = {}): UseChatHelpers & {
+  experimental_addToolResult: ({
+    toolCallId,
+    result,
+  }: {
+    toolCallId: string;
+    result: any;
+  }) => void;
+} {
   // Generate a unique id for the chat if not provided.
   const hookId = useId();
   const idKey = id ?? hookId;
@@ -324,6 +357,23 @@ export function useChat({
       } finally {
         mutateLoading(false);
       }
+
+      // auto-submit when all tool calls in the last assistant message have results:
+      const messages = messagesRef.current;
+      const lastMessage = messages[messages.length - 1];
+      if (
+        // ensure there is a last message:
+        lastMessage != null &&
+        // check if the feature is enabled:
+        experimental_maxAutomaticRoundtrips > 0 &&
+        // check that roundtrip is possible:
+        isAssistantMessageWithCompletedToolCalls(lastMessage) &&
+        // limit the number of automatic roundtrips:
+        countTrailingAssistantMessages(messages) <=
+          experimental_maxAutomaticRoundtrips
+      ) {
+        await triggerRequest({ messages });
+      }
     },
     [
       mutate,
@@ -340,6 +390,7 @@ export function useChat({
       sendExtraMessageFields,
       experimental_onFunctionCall,
       experimental_onToolCall,
+      experimental_maxAutomaticRoundtrips,
       messagesRef,
       abortControllerRef,
       generateId,
@@ -480,5 +531,65 @@ export function useChat({
     handleSubmit,
     isLoading,
     data: streamData,
+    experimental_addToolResult: ({
+      toolCallId,
+      result,
+    }: {
+      toolCallId: string;
+      result: any;
+    }) => {
+      const updatedMessages = messagesRef.current.map((message, index, arr) =>
+        // update the tool calls in the last assistant message:
+        index === arr.length - 1 &&
+        message.role === 'assistant' &&
+        message.toolInvocations
+          ? {
+              ...message,
+              toolInvocations: message.toolInvocations.map(toolInvocation =>
+                toolInvocation.toolCallId === toolCallId
+                  ? { ...toolInvocation, result }
+                  : toolInvocation,
+              ),
+            }
+          : message,
+      );
+
+      mutate(updatedMessages, false);
+
+      // auto-submit when all tool calls in the last assistant message have results:
+      const lastMessage = updatedMessages[updatedMessages.length - 1];
+      if (isAssistantMessageWithCompletedToolCalls(lastMessage)) {
+        triggerRequest({ messages: updatedMessages });
+      }
+    },
   };
+}
+
+/**
+Check if the message is an assistant message with completed tool calls. 
+The message must have at least one tool invocation and all tool invocations
+must have a result.
+ */
+function isAssistantMessageWithCompletedToolCalls(message: Message) {
+  return (
+    message.role === 'assistant' &&
+    message.toolInvocations &&
+    message.toolInvocations.length > 0 &&
+    message.toolInvocations.every(toolInvocation => 'result' in toolInvocation)
+  );
+}
+
+/**
+Returns the number of trailing assistant messages in the array.
+ */
+function countTrailingAssistantMessages(messages: Message[]) {
+  let count = 0;
+  for (let i = messages.length - 1; i >= 0; i--) {
+    if (messages[i].role === 'assistant') {
+      count++;
+    } else {
+      break;
+    }
+  }
+  return count;
 }