+++ packages/openai/src/openai-completion-language-model.test.ts
@@ -38,7 +38,11 @@ const TEST_LOGPROBS = {
   ] as Record<string, number>[],
 };
 
-const provider = createOpenAI({ apiKey: 'test-api-key' });
+const provider = createOpenAI({
+  apiKey: 'test-api-key',
+  compatibility: 'strict',
+});
+
 const model = provider.completion('gpt-3.5-turbo-instruct');
 
 describe('doGenerate', () => {
@@ -133,7 +137,7 @@ describe('doGenerate', () => {
     );
   });
 
-  it('should finish reason', async () => {
+  it('should extract finish reason', async () => {
     prepareJsonResponse({
       content: '',
       finish_reason: 'stop',
@@ -348,6 +352,7 @@ describe('doStream', () => {
 
     expect(await server.getRequestBodyJson()).toStrictEqual({
       stream: true,
+      stream_options: { include_usage: true },
       model: 'gpt-3.5-turbo-instruct',
       prompt: 'Hello',
     });